+++ src/app/rankings/page.tsx
@@ -4,10 +4,10 @@ import { getCompetitionList, getDisciplineList } from "@lib/database/dbutils";
 import RankingsClient from "./RankingsClient";
 
 // Types
-import { GenericStringIndexWithDate, GenericStringIndex } from "@/app/type/generic";
+import { GenericStringIndexWithDate } from "@/app/type/generic";
 
 const RankingsPage = async () => {
-  const [competitionList, disciplinesList]: [GenericStringIndexWithDate[], GenericStringIndexWithDate[]] = await Promise.all([
+  const [competitionList, disciplineList]: [GenericStringIndexWithDate[], GenericStringIndexWithDate[]] = await Promise.all([
     getCompetitionList(["id", "name", "city", "type"]),
     getDisciplineList(),
   ]);
@@ -15,7 +15,7 @@ const RankingsPage = async () => {
   return (
     <div className="page page-rankings">
       <h2 className="page page-title">Classements</h2>
-      <RankingsClient competitionList={competitionList} disciplinesList={disciplinesList} />
+      <RankingsClient competitionList={competitionList} disciplineList={disciplineList} />
     </div>
   );
 };