+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix any typescript and eslint error, rename variables.
 - Refacto Results and Competitions, use SSR, finish results refacto with SQL data.
 - Use Layout for results page, split components for ssr, use router for results pages.
+- Fix ranking sort, fix syntax and typescript errors.
 
 ## v0.12.0 - 2025-06-22