+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Refacto Results and Competitions, use SSR, finish results refacto with SQL data.
 - Use Layout for results page, split components for ssr, use router for results pages.
 - Fix ranking sort, fix syntax and typescript errors.
-
+- Add "prisma generate" to build.
 ## v0.12.0 - 2025-06-22
 
 ### Added